MiGz 2.0.beta-1: Rewrite MiGz to use ForkJoinPool
Version 2 of MiGz uses ForkJoinPools, with the primary benefit that the commonPool() (or another, existing ForkJoinPool) may be used, which can help avoid the number of concurrently executing threads exceeding the number of logical cores (when multiple instances of MiGz streams or used together, or when there is other parallelized work being done).  MiGz is also much more efficient in how it uses its buffers, allowing as much as twice the previous amount of buffering capacity with roughly the same amount of memory consumed.

Because version 2 marks a substantial departure from MiGz 1.x, we consider this "beta" as it is hard to thoroughly test for events like exceptions from the underlying stream or close() calls at arbitrary times using unit tests.  We are awaiting more experience using this class in real-world tasks before removing the "beta" designation.